[feature] Welcome banner
Reason: Firebase websockets loads app data asynchronously, and AFAIK
there is no hook for determining if no data exists or if data has not
been fetched yet.

In the future, this can be refined to wait for initial fetch, if there
is a way. E.g. banner can listen for posts, and onChange, hide itself.
But simple solution seems to work fine for now.

Changes:
* Refactor header text into <Logo/>
* Move <Clock/> to feature
* Add <Welcome/> feature